refactor geometry module
