+++ survey_cad/src/geometry/point3.rs
@@ -1,0 +1,16 @@
+//! Basic 3D point type used throughout the crate.
+
+/// Representation of a 3D point.
+#[derive(Debug, Clone, Copy, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct Point3 {
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+}
+
+impl Point3 {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
+        Self { x, y, z }
+    }
+}
+
